add bottom navigation and rebuild architecture app
