+++ .idea/misc.xml
@@ -3,32 +3,44 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Users/erlin/AndroidStudioProjects/traveloka/app/src/main/res/layout/activity_add_campaign.xml" value="0.17074275362318841" />
-        <entry key="..\:/Users/erlin/AndroidStudioProjects/traveloka/app/src/main/res/layout/activity_add_product.xml" value="0.17074275362318841" />
-        <entry key="..\:/Users/erlin/AndroidStudioProjects/traveloka/app/src/main/res/layout/activity_add_promo.xml" value="0.17074275362318841" />
-        <entry key="..\:/Users/erlin/AndroidStudioProjects/traveloka/app/src/main/res/layout/activity_campaign.xml" value="0.17074275362318841" />
-        <entry key="..\:/Users/erlin/AndroidStudioProjects/traveloka/app/src/main/res/layout/activity_detail_customer.xml" value="0.17074275362318841" />
-        <entry key="..\:/Users/erlin/AndroidStudioProjects/traveloka/app/src/main/res/layout/activity_list_customer.xml" value="0.17074275362318841" />
-        <entry key="..\:/Users/erlin/AndroidStudioProjects/traveloka/app/src/main/res/layout/item_voucher.xml" value="0.17074275362318841" />
-        <entry key="..\:/Users/erlin/AndroidStudioProjects/traveloka/app/src/main/res/layout/layout_list_customer.xml" value="0.17074275362318841" />
         <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.262" />
         <entry key="app/src/main/res/drawable/box.xml" value="0.2195" />
         <entry key="app/src/main/res/drawable/ic_baseline_add_24.xml" value="0.316" />
+        <entry key="app/src/main/res/drawable/ic_baseline_calendar_month_24.xml" value="0.285" />
+        <entry key="app/src/main/res/drawable/ic_baseline_campaign_24.xml" value="0.285" />
+        <entry key="app/src/main/res/drawable/ic_baseline_customers_24.xml" value="0.285" />
         <entry key="app/src/main/res/drawable/ic_baseline_image_24.xml" value="0.262" />
+        <entry key="app/src/main/res/drawable/ic_baseline_local_activity_24.xml" value="0.285" />
+        <entry key="app/src/main/res/drawable/ic_baseline_people_24.xml" value="0.285" />
+        <entry key="app/src/main/res/drawable/ic_baseline_product_24.xml" value="0.285" />
+        <entry key="app/src/main/res/drawable/ic_baseline_widgets_24.xml" value="0.285" />
+        <entry key="app/src/main/res/drawable/ic_dashboard_black_24dp.xml" value="0.2355" />
+        <entry key="app/src/main/res/drawable/ic_home_black_24dp.xml" value="0.2355" />
         <entry key="app/src/main/res/drawable/ic_launcher_background.xml" value="0.262" />
         <entry key="app/src/main/res/layout/activity_add_campaign.xml" value="0.30416666666666664" />
         <entry key="app/src/main/res/layout/activity_add_product.xml" value="0.3333333333333333" />
         <entry key="app/src/main/res/layout/activity_add_promo.xml" value="0.30416666666666664" />
         <entry key="app/src/main/res/layout/activity_campaign.xml" value="0.30416666666666664" />
+        <entry key="app/src/main/res/layout/activity_detail_customer.xml" value="0.30416666666666664" />
         <entry key="app/src/main/res/layout/activity_detail_product.xml" value="0.5" />
         <entry key="app/src/main/res/layout/activity_detail_voucher.xml" value="0.30416666666666664" />
         <entry key="app/src/main/res/layout/activity_forecast.xml" value="0.4466666666666667" />
         <entry key="app/src/main/res/layout/activity_list_customer.xml" value="0.30416666666666664" />
         <entry key="app/src/main/res/layout/activity_list_promo.xml" value="0.30416666666666664" />
         <entry key="app/src/main/res/layout/activity_list_voucher.xml" value="0.30416666666666664" />
-        <entry key="app/src/main/res/layout/activity_main.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/activity_main.xml" value="0.3333333333333333" />
+        <entry key="app/src/main/res/layout/activity_product_list.xml" value="0.30416666666666664" />
+        <entry key="app/src/main/res/layout/fragment_campaign.xml" value="0.30416666666666664" />
+        <entry key="app/src/main/res/layout/fragment_customer.xml" value="0.30416666666666664" />
+        <entry key="app/src/main/res/layout/fragment_dashboard.xml" value="0.30416666666666664" />
+        <entry key="app/src/main/res/layout/fragment_home.xml" value="0.30416666666666664" />
+        <entry key="app/src/main/res/layout/fragment_list_voucher.xml" value="0.30416666666666664" />
+        <entry key="app/src/main/res/layout/fragment_notifications.xml" value="0.30416666666666664" />
+        <entry key="app/src/main/res/layout/fragment_product.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/fragment_voucher.xml" value="0.1" />
         <entry key="app/src/main/res/layout/item_row_campaign.xml" value="0.6" />
-        <entry key="app/src/main/res/layout/item_row_product.xml" value="0.22" />
+        <entry key="app/src/main/res/layout/item_row_product.xml" value="0.4466666666666667" />
+        <entry key="app/src/main/res/menu/bottom_nav_menu.xml" value="0.30416666666666664" />
         <entry key="app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.2195" />
       </map>
     </option>
