+++ app/src/main/java/com/traveloka/bestpriceapp/ui/MainActivity.kt
@@ -28,7 +28,7 @@ class MainActivity : AppCompatActivity() {
         // menu should be considered as top level destinations.
         val appBarConfiguration = AppBarConfiguration(
             setOf(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_notifications
+                R.id.navigation_product, R.id.navigation_voucher, R.id.navigation_customer, R.id.navigation_campaign
             )
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
