+++ app/src/main/java/com/traveloka/bestpriceapp/ui/campaign/activity/AddCampaignActivity.kt
@@ -1,12 +1,12 @@
-package com.traveloka.bestpriceapp.ui.promo.activity
+package com.traveloka.bestpriceapp.ui.campaign.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.traveloka.bestpriceapp.R
 
-class ForecastActivity : AppCompatActivity() {
+class AddCampaignActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_forecast)
+        setContentView(R.layout.activity_add_campaign)
     }
 }
