+++ app/src/main/java/com/traveloka/bestpriceapp/ui/campaign/activity/ForecastActivity.kt
@@ -1,12 +1,12 @@
-package com.traveloka.bestpriceapp.ui.product.activity
+package com.traveloka.bestpriceapp.ui.campaign.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.traveloka.bestpriceapp.R
 
-class ProductListActivity : AppCompatActivity() {
+class ForecastActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_product_list)
+        setContentView(R.layout.activity_forecast)
     }
 }
