+++ app/src/main/java/com/traveloka/bestpriceapp/ui/campaign/activity/ListPromoActivity.kt
@@ -1,12 +1,12 @@
-package com.traveloka.bestpriceapp.ui.customer.activity
+package com.traveloka.bestpriceapp.ui.campaign.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.traveloka.bestpriceapp.R
 
-class ListCustomerActivity : AppCompatActivity() {
+class ListPromoActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_list_customer)
+        setContentView(R.layout.activity_list_promo)
     }
 }
