+++ app/src/main/java/com/traveloka/bestpriceapp/ui/campaign/fragment/CampaignFragment.kt
@@ -1,4 +1,4 @@
-package com.traveloka.bestpriceapp.ui.ui.dashboard
+package com.traveloka.bestpriceapp.ui.campaign.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,12 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.traveloka.bestpriceapp.ui.databinding.FragmentDashboardBinding
+import com.traveloka.bestpriceapp.databinding.FragmentCampaignBinding
+import com.traveloka.bestpriceapp.ui.campaign.viewmodel.CampaignViewModel
 
-class DashboardFragment : Fragment() {
+class CampaignFragment : Fragment() {
 
-    private var _binding: FragmentDashboardBinding? = null
+    private var _binding: FragmentCampaignBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,16 +23,16 @@ class DashboardFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val dashboardViewModel =
-            ViewModelProvider(this).get(DashboardViewModel::class.java)
+        val notificationsViewModel =
+            ViewModelProvider(this).get(CampaignViewModel::class.java)
 
-        _binding = FragmentDashboardBinding.inflate(inflater, container, false)
+        _binding = FragmentCampaignBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textDashboard
-        dashboardViewModel.text.observe(viewLifecycleOwner) {
-            textView.text = it
-        }
+//        val textView: TextView = binding.textNotifications
+//        notificationsViewModel.text.observe(viewLifecycleOwner) {
+//            textView.text = it
+//        }
         return root
     }
