+++ app/src/main/java/com/traveloka/bestpriceapp/ui/campaign/viewmodel/CampaignViewModel.kt
@@ -1,13 +1,13 @@
-package com.traveloka.bestpriceapp.ui.ui.home
+package com.traveloka.bestpriceapp.ui.campaign.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class CampaignViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is Campaign Fragment"
     }
     val text: LiveData<String> = _text
 }
