+++ app/src/main/java/com/traveloka/bestpriceapp/ui/customer/fragment/CustomerFragment.kt
@@ -1,4 +1,4 @@
-package com.traveloka.bestpriceapp.ui.ui.notifications
+package com.traveloka.bestpriceapp.ui.customer.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,12 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.traveloka.bestpriceapp.databinding.FragmentNotificationsBinding
+import com.traveloka.bestpriceapp.databinding.FragmentCustomerBinding
+import com.traveloka.bestpriceapp.ui.customer.viewmodel.CustomerViewModel
 
-class NotificationsFragment : Fragment() {
+class CustomerFragment : Fragment() {
 
-    private var _binding: FragmentNotificationsBinding? = null
+    private var _binding: FragmentCustomerBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,16 +23,16 @@ class NotificationsFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val notificationsViewModel =
-            ViewModelProvider(this).get(NotificationsViewModel::class.java)
+        val dashboardViewModel =
+            ViewModelProvider(this).get(CustomerViewModel::class.java)
 
-        _binding = FragmentNotificationsBinding.inflate(inflater, container, false)
+        _binding = FragmentCustomerBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textNotifications
-        notificationsViewModel.text.observe(viewLifecycleOwner) {
-            textView.text = it
-        }
+//        val textView: TextView = binding.textDashboard
+//        dashboardViewModel.text.observe(viewLifecycleOwner) {
+//            textView.text = it
+//        }
         return root
     }
