+++ app/src/main/java/com/traveloka/bestpriceapp/ui/customer/viewmodel/CustomerViewModel.kt
@@ -1,13 +1,13 @@
-package com.traveloka.bestpriceapp.ui.ui.notifications
+package com.traveloka.bestpriceapp.ui.customer.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class CustomerViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is Customer Fragment"
     }
     val text: LiveData<String> = _text
 }
