+++ app/src/main/java/com/traveloka/bestpriceapp/ui/product/fragment/ProductFragment.kt
@@ -1,4 +1,4 @@
-package com.traveloka.bestpriceapp.ui.ui.home
+package com.traveloka.bestpriceapp.ui.product.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,12 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.traveloka.bestpriceapp.databinding.FragmentHomeBinding
+import com.traveloka.bestpriceapp.databinding.FragmentProductBinding
+import com.traveloka.bestpriceapp.ui.product.viewmodel.ProductViewModel
 
-class HomeFragment : Fragment() {
+class ProductFragment : Fragment() {
 
-    private var _binding: FragmentHomeBinding? = null
+    private var _binding: FragmentProductBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -23,15 +24,15 @@ class HomeFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         val homeViewModel =
-            ViewModelProvider(this).get(HomeViewModel::class.java)
+            ViewModelProvider(this).get(ProductViewModel::class.java)
 
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentProductBinding.inflate(inflater, container, false)
         val root: View = binding.root
-
-        val textView: TextView = binding.textHome
-        homeViewModel.text.observe(viewLifecycleOwner) {
-            textView.text = it
-        }
+//
+//        val textView: TextView = binding.textHome
+//        homeViewModel.text.observe(viewLifecycleOwner) {
+//            textView.text = it
+//        }
         return root
     }
