+++ app/src/main/java/com/traveloka/bestpriceapp/ui/product/viewmodel/ProductViewModel.kt
@@ -1,13 +1,13 @@
-package com.traveloka.bestpriceapp.ui.ui.dashboard
+package com.traveloka.bestpriceapp.ui.product.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class ProductViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is Product Fragment"
     }
     val text: LiveData<String> = _text
 }
