--- app/src/main/java/com/traveloka/bestpriceapp/ui/promo/activity/AddPromoActivity.kt
@@ -1,12 +1,0 @@
-package com.traveloka.bestpriceapp.ui.promo.activity
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.traveloka.bestpriceapp.R
-
-class AddPromoActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_add_promo)
-    }
-}
