+++ app/src/main/java/com/traveloka/bestpriceapp/ui/voucher/fragment/VoucherFragment.kt
@@ -35,7 +35,7 @@ class ListVoucherFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_list_voucher, container, false)
+        return inflater.inflate(R.layout.fragment_voucher, container, false)
     }
 
     companion object {
