+++ app/src/main/res/layout/activity_add_campaign.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.promo.activity.AddCampaignActivity">
+    tools:context=".ui.campaign.activity.AddCampaignActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
