+++ app/src/main/res/layout/activity_add_promo.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.promo.activity.AddPromoActivity">
+    tools:context=".ui.campaign.activity.AddPromoActivity">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/textInputLayout"
