+++ app/src/main/res/layout/activity_forecast.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.promo.activity.ForecastActivity">
+    tools:context=".ui.campaign.activity.ForecastActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
