+++ app/src/main/res/layout/activity_list_promo.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.promo.activity.ListPromoActivity">
+    tools:context=".ui.campaign.activity.ListPromoActivity">
 
     <TextView
         android:id="@+id/textView3"
