+++ app/src/main/res/layout/fragment_campaign.xml
