+++ app/src/main/res/layout/item_row_product.xml
@@ -36,6 +36,7 @@
             android:layout_toEndOf="@id/img_item_photo"
             android:textSize="20sp"
             android:textStyle="normal"
+            android:layout_marginStart="10dp"
             card_view:layout_constraintBottom_toTopOf="@+id/tv_item_username"
             card_view:layout_constraintEnd_toEndOf="parent"
             card_view:layout_constraintHorizontal_bias="0.5"
@@ -46,7 +47,7 @@
         <TextView
             android:id="@+id/tv_item_username"
             style="@style/itemUsername"
-            android:layout_width="227dp"
+            android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:textSize="12sp"
             card_view:layout_constraintBottom_toTopOf="@+id/tv_campaign"
@@ -57,7 +58,7 @@
 
         <TextView
             android:id="@+id/textView"
-            android:layout_width="227dp"
+            android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:text="Rp. 20000"
             android:textSize="24sp"
